@@ -4,24 +4,24 @@ import '../constants.dart';
 
 class DefaultButton extends StatelessWidget {
   const DefaultButton({
-    Key key,
-    this.imageSrc,
-    this.text,
-    this.press,
+    Key? key,
+    required this.imageSrc,
+    required this.text,
+    required this.press,
   }) : super(key: key);
 
   final String imageSrc, text;
-  final Function press;
+  final VoidCallback press;
 
   @override
   Widget build(BuildContext context) {
-    return FlatButton(
-      padding: EdgeInsets.symmetric(
-        vertical: kDefaultPadding,
-        horizontal: kDefaultPadding * 2.5,
-      ),
-      shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(50)),
-      color: Color(0xFFE8F0F9),
+    return ElevatedButton(
+      // padding: EdgeInsets.symmetric(
+      //   vertical: kDefaultPadding,
+      //   horizontal: kDefaultPadding * 2.5,
+      // ),
+      // shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(50)),
+      // color: Color(0xFFE8F0F9),
       onPressed: press,
       child: Row(
         children: [
