@@ -4,10 +4,10 @@ import '../constants.dart';
 
 class SectionTitle extends StatelessWidget {
   const SectionTitle({
-    Key key,
-    this.title,
-    this.subTitle,
-    this.color,
+    Key? key,
+    required this.title,
+    required this.subTitle,
+    required this.color,
   }) : super(key: key);
 
   final String title, subTitle;
@@ -46,7 +46,7 @@ class SectionTitle extends StatelessWidget {
                 title,
                 style: Theme.of(context)
                     .textTheme
-                    .headline2
+                    .headlineMedium!
                     .copyWith(fontWeight: FontWeight.bold, color: Colors.black),
               )
             ],
