@@ -5,7 +5,12 @@ class Feedback {
   final int id;
   final Color color;
 
-  Feedback({this.name, this.review, this.userPic, this.id, this.color});
+  Feedback(
+      {required this.name,
+      required this.review,
+      required this.userPic,
+      required this.id,
+      required this.color});
 }
 
 // List of demo feedbacks
